requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
openpyxl==3.1.5
packaging==26.0
pillow==12.1.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytesseract==0.3.13
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
